chore: code review -- tweak type definitions
